refactor: renamed package names to not include camel casing
