+++ go/test/endtoend/vtorc/gracefultakeover/graceful_takeover_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package gracefulTakeover
+package gracefultakeover
 
 import (
 	"fmt"
