+++ go/test/endtoend/vtorc/primaryfailure/main_test.go
@@ -14,15 +14,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package gracefulTakeover
+package primaryfailure
 
 import (
 	"fmt"
 	"os"
 	"testing"
 
-	"vitess.io/vitess/go/test/endtoend/cluster"
 	"vitess.io/vitess/go/test/endtoend/vtorc/utils"
+
+	"vitess.io/vitess/go/test/endtoend/cluster"
 )
 
 var clusterInfo *utils.VtOrcClusterInfo
@@ -32,9 +33,15 @@ func TestMain(m *testing.M) {
 	var cellInfos []*utils.CellInfo
 	cellInfos = append(cellInfos, &utils.CellInfo{
 		CellName:    utils.Cell1,
-		NumReplicas: 6,
+		NumReplicas: 12,
 		NumRdonly:   2,
 		UIDBase:     100,
+	})
+	cellInfos = append(cellInfos, &utils.CellInfo{
+		CellName:    utils.Cell2,
+		NumReplicas: 2,
+		NumRdonly:   0,
+		UIDBase:     200,
 	})
 
 	exitcode, err := func() (int, error) {
